requirements.txt
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_full_version <= "3.11.2"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
coverage==7.2.7 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.10" and python_version < "4.0"
django-environ==0.10.0 ; python_version >= "3.10" and python_version < "4"
django-prometheus==2.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-redis==5.3.0 ; python_version >= "3.10" and python_version < "4.0"
django-stubs-ext==4.2.2 ; python_version >= "3.10" and python_version < "4.0"
django-stubs==4.2.3 ; python_version >= "3.10" and python_version < "4.0"
django==4.2.4 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.2.2 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-stubs==3.14.2 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.26.4 ; python_version >= "3.10" and python_version < "4.0"
factory-boy==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
faker==19.3.0 ; python_version >= "3.10" and python_version < "4.0"
flake8==6.1.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.56.2 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.56.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.18.4 ; python_version >= "3.10" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
mypy==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.18.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger-proto-grpc==1.19.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger-thrift==1.19.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger==1.19.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-dbapi==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-django==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-logging==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-psycopg2==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-wsgi==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.18.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.39b0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.17.1 ; python_version >= "3.10" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.10" and python_version < "4.0"
pycodestyle==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.30.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
result==0.10.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.9.2 ; python_version >= "3.10" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
thrift==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
types-pytz==2023.3.0.0 ; python_version >= "3.10" and python_version < "4.0"
types-pyyaml==6.0.12.11 ; python_version >= "3.10" and python_version < "4.0"
types-requests==2.31.0.2 ; python_version >= "3.10" and python_version < "4.0"
types-urllib3==1.26.25.14 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.16.2 ; python_version >= "3.10" and python_version < "4.0"
